Fix further bug in loading mutants from disk.
This change ensures that the correct subject root path is used
when a mutant is loaded from disk. Previously, we set the root
path to be the ".mutants" directory, but this is clearly not the
correct root path for a subject. We now instead use the existing
subject set for the project to find a subject that has the correct
name and relative path for each mutant stored on disk.