Select tests for subject when expressions overlap.
The previous implementation was too strict and would only select a test
for a subject if the test's expression matched the subject's name. We
now select tests whose expression appears at the start of the subject's
name. So for example, tests for "Min" are selected for the "Min::Batch"
subject. For some projects, this is overkill as the "Min" tests won't be
relevant for "Min::Batch", but in other projects it is necessary as
tests are written against a parent module rather than against a nested
subject ("Min" here rather than "Min::Batch").