Develop more sophisticated test filtering.
Previously, tests were selected only if the subject's name began with
the test's expression (which was the first word of the RSpec full
description). However, this had a few downsides:

1) Tests that had expressions of the form Class#method would never be
selected.
2) Tests that had expressions that do not relate to a subject (e.g.,
feature tests) would never be selected.

This change adds a new test selection strategy that:

1) Always selects a test if its full description does not start with any
known subject.
2) Selects a test if its full description starts with the subject's
name.
3) Otherwise, does not select a test.

This means that generic tests are run for every subject, and that
targeted tests are run for the relevant subject (even when the test's
name is narrower than the subject, e.g., Max#run is selected for Max).